refactor: refactor googlenet
