+++ models/GoogLeNet/__init__.py
@@ -1,1 +1,1 @@
-from .model import GoogLeNet
+from .lightning_model import *
